root_agent/requirements.txt
google-cloud-aiplatform[agent_engines,adk]
google-cloud-storage
google-genai